Use defoverridable and add reflection to Norm.Contract
This is arguably simpler impelementation where we do less
metaprogramming by leveraging defoverridable.

Instead of overriding Kernel.def/2, we let the function be defined as
usual and we later simply defoverridable it with the contract conforming
function.

Additionally we add contract reflection, that is for each contract there
would be a `__contract__({fun, arity})` clause defined on the module.
This could be later used as a building block for things like:

    iex> contract = Foo.__contract__({:add, 2})
    iex> Norm.Contract.conform!(contract, {Bar, :bar, ["1", 2]})
    ** (Norm.MismatchError) Could not conform input:
    val: 42 fails: is_binary()
    iex> Norm.Contract.conform!(contract, {Bar, :bar, [1, 2]})
    3

but currently it's just an implementation detail. The reflection
function is not yet documented and is considered private API.